--- Source/DataIngestion/NCReadFile.H
@@ -1,8 +1,0 @@
-#ifndef _NCREADFILE_H_
-#define _NCPREADILE_H_
-
-void
-ERF::ncrestart(istream& is)
-{};
-
-#endif
