+++ Source/Make.package
@@ -1,5 +1,6 @@
 CEXE_sources += main.cpp
 CEXE_sources += ERF.cpp
+CEXE_sources += ERF_init.cpp
 CEXE_headers += ERF.H
 CEXE_headers += prob_common.H
